Bug fixes and optimizations.

+++ sha1.cpp
@@ -1,6 +1,7 @@
 #include "headers/sha1.h"
 #include <string.h>
 
+
 uint cycle_shift_left(uint val, int bit_count) {
     return (val << bit_count | val >> (32-bit_count));
 }
@@ -13,17 +14,17 @@ uint bring_to_human_view(uint val) {
 }
 
 uint* sha1(const char* message, uint msize_bytes) {
-    //инициализация
+    //init
     uint A = H[0];
     uint B = H[1];
     uint C = H[2];
     uint D = H[3];
     uint E = H[4];
     
-    // подсчет целого числа блоков
+    // total number of blocks
     uint totalBlockCount = msize_bytes  / one_block_size_bytes;
     
-    // подсчет, сколько байт нужно, чтобы дополнить последний блок
+    // how many bytes are required to supplement the last block
     uint needAdditionalBytes = 
         one_block_size_bytes - (msize_bytes - totalBlockCount * one_block_size_bytes);
         
@@ -34,34 +35,34 @@ uint* sha1(const char* message, uint msize_bytes) {
         totalBlockCount += 1;
     }
     
-    // размер дополненного по всем правилам сообщения
+    // size of a corretly supplemented message
     uint extendedMessageSize = msize_bytes + needAdditionalBytes;
     
-    // выделяем новый буфер и копируем в него исходный
+    // allocating a new buffer and copying the original block into it
     unsigned char* newMessage = new unsigned char[extendedMessageSize];
     memcpy(newMessage, message, msize_bytes);
     
-    // первый бит ставим '1', остальные обнуляем
+    // first bit '1', others '0'
     newMessage[msize_bytes] = 0x80;    
     memset(newMessage + msize_bytes + 1, 0, needAdditionalBytes - 1);
     
-    // задаем длину исходного сообщения в битах
+    // original message length in bits
     uint* ptr_to_size = (uint*) (newMessage + extendedMessageSize - 4); 
     *ptr_to_size =  bring_to_human_view(msize_bytes * 8);
     
     ExpendBlock exp_block;
-    //раунды поехали
-    for(int i=0; i<totalBlockCount; i++) {
+    // rounds
+    for(size_t i=0; i<totalBlockCount; i++) {
         
-        // берем текущий блок и дополняем его
+        // supplementing the current block
         unsigned char* cur_p = newMessage + one_block_size_bytes * i;
         Block block = (Block) cur_p;
         
-        // первые 16 4байтовых чисел
+        // first 16 4-byte numbers
         for (int j=0; j<one_block_size_uints; j++) {
             exp_block[j] = bring_to_human_view(block[j]);
         }
-        // следующие 64...
+        // next 64...
         for (int j=one_block_size_uints; j < block_expend_size_uints; j++) {
             exp_block[j] = 
                 exp_block[j-3] ^ 
@@ -71,18 +72,18 @@ uint* sha1(const char* message, uint msize_bytes) {
             exp_block[j] = cycle_shift_left(exp_block[j], 1);
         }
             
-        // инициализация 
+        // init 
         uint a = H[0];
         uint b = H[1];
         uint c = H[2];
         uint d = H[3];
         uint e = H[4];
 
-        // пересчитываем
+        // recalculating
         for(int j=0;j < block_expend_size_uints; j++) {
             uint f;
             uint k;
-            // в зависимости от раунда считаем по-разному
+            // counting differently depending on round
             if (j < 20) {
                 f = (b & c) | ((~b) & d);
                 k = 0x5A827999;
@@ -97,7 +98,7 @@ uint* sha1(const char* message, uint msize_bytes) {
                 k = 0xCA62C1D6;
             }
         
-            // перемешивание
+            // shuffling
             uint temp = cycle_shift_left(a,5) + f + e + k + exp_block[j];
             e = d;
             d = c;
@@ -105,7 +106,7 @@ uint* sha1(const char* message, uint msize_bytes) {
             b = a;
             a = temp;   
         }
-        // пересчитываем
+        // recalculating
         A = A + a;
         B = B + b;
         C = C + c;
@@ -113,7 +114,7 @@ uint* sha1(const char* message, uint msize_bytes) {
         E = E + e;
     }
     
-    // A,B,C,D,E являются выходными 32б составляющими посчитанного хэша
+    // A,B,C,D,E are the out 32-bit components of the calculated hash
     uint* digest = new uint[5];
     digest[0] = A;
     digest[1] = B;
@@ -121,7 +122,6 @@ uint* sha1(const char* message, uint msize_bytes) {
     digest[3] = D;
     digest[4] = E;
     
-    // чистим за собой
     delete[] newMessage;
     return digest;
 }
